app_module/backend/requirements.txt
fastapi==0.127.1
uvicorn[standard]==0.40.0

motor==3.7.1
pydantic[email]==2.7.4
python-dotenv==1.0.1

celery==5.6.0
redis==7.1.0

passlib[bcrypt]==1.7.4
bcrypt==3.2.2

python-jose[cryptography]==3.3.0

httpx